Wind, temperature and cloud over the map, from one forecast grid
He asked to see whether the map is in sunshine or cloud, what the
temperature is doing, and above all the wind — his words: wind especially
needs taking into account when planning. All toggleable, so the map is a
map again when none of it is wanted.

One request does all of it. A five-by-five grid of points over the
viewport brings back temperature, rain, cloud, wind speed and wind
direction for every hour from yesterday to tomorrow, so switching a
second layer on costs nothing, and dragging the timeline costs nothing at
all. It also fills out the ten-hour scrubber the radar could only cover
two and a half hours of.

Wind is the one he emphasised, so it gets arrows rather than a wash:
coloured and lengthened by speed, grey a breath through to red where an
exposed ridge stops being a good idea. Each arrow flies the way the wind
is going, while the reading beside the timeline names the direction it
comes from — both are needed, and stating only one of them is how a wind
arrow misleads. The walk brief now reports direction and gusts too, with
a plain sentence about what that speed means on open ground, because a
number with no direction cannot answer "is it behind me".

Temperature and cloud are colour washes, and only one wash is drawn at a
time — two over each other say nothing legible. Forecast rain takes the
map wherever the radar cannot see, then temperature, then cloud. Cloud
marks clear sky in faint gold rather than leaving it blank: "sunny" and
"no data" must not look the same.

Held apart from the radar throughout. These fields are a model — the
forecast's opinion about the sky, a few kilometres between readings —
where radar is a measurement of rain that fell. The frame label says
which, the Settings text says which, and the timeline never lays a
forecast hour on top of an observed sweep.

The colour ramps and the smoothing are plain arithmetic with no Android
in them, and tested: rain is never invisible where there is rain, heavier
is never more transparent than lighter, clear sky reads warm and overcast
neutral, and a reading the forecast omitted is carried by its neighbours
rather than dragging a quarter of the map to zero.

+++ phone/src/test/java/com/jollydoddger/waymark/TimelineTest.kt
@@ -62,7 +62,7 @@ class TimelineTest {
     @Test
     fun withNoRadarAtAllTheForecastStillFillsTheTimeline() {
         val merged = Timeline.merge(emptyList(), hours(-5, 5), now)
-        assertEquals(11, merged.size)
+        assertEquals(11L, merged.size.toLong())
         merged.forEach { assertNull(it.radarPath) }
         assertEquals("forecast", merged.first().kind)
     }
